+++ beaker/data_model/dataset.py
@@ -26,11 +26,13 @@ class Dataset(BaseModel):
     id: str
     name: Optional[str] = None
     full_name: Optional[str] = None
+    description: Optional[str] = None
     owner: Account
     author: Account
     created: datetime
     committed: Optional[datetime] = None
     workspace_ref: WorkspaceRef
+    source_execution: Optional[str] = None
     storage: Optional[DatasetStorage] = None
 
     @property