+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `Beaker.experiment.latest_job()` method.
+
 ## [v0.15.1](https://github.com/allenai/beaker-py/releases/tag/v0.15.1) - 2022-05-06
 
 ### Added