+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `Beaker.experiment.latest_job()` method.
 - Added `strict: bool` argument to `Beaker.(experiment|job).(wait_for|as_completed)` methods.
+- Added `description` and `source_execution` fields to `Dataset`.
+- Added `description` field to `Experiment`.
+- Added `description` field to `Image`.
+- Added `description` field to `Workspace`.
 
 ## [v0.15.1](https://github.com/allenai/beaker-py/releases/tag/v0.15.1) - 2022-05-06